# openapi.yaml
openapi: 3.1.0
info:
  title: Collab Canvas Aid API
  version: 0.1.0
  description: |
    HTTP and WebSocket contract expected by the frontend in frontend/.

    The checked-in frontend currently uses an in-browser MockBackend rather than
    an HTTP client. This specification maps the Services interface and the API
    surface described by frontend/README.md to a backend contract. Authentication
    is delegated to an external identity provider; the API receives its resulting
    bearer token.
servers:
  - url: https://api.example.com
    description: Replace with the deployed API origin.
...
paths:
  /v1/auth/me:
    get:
      tags: [Authentication]
      operationId: getCurrentUser
      summary: Get the signed-in user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current authenticated user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/auth/signout:
    post:
      tags: [Authentication]
      operationId: signOut
      summary: Sign out the current user
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Sign-out completed.
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/auth/token:
    post:
      tags: [Authentication]
      operationId: issueToken
      summary: Exchange a password for a bearer token
      description: Development/MVP login endpoint for the in-memory backend.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '200':
          description: Bearer access token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /v1/sessions/{id}/canvas:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    get:
      tags: [Canvas]
      operationId: getCanvasSnapshot
      summary: Get the authoritative canvas snapshot
      description: Requires session access through a user token or participant collaboration credential.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current canvas document and operation cursor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CanvasSnapshot'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/sessions/{id}/canvas/operations:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    post:
      tags: [Canvas]
      operationId: submitCanvasOperations
      summary: Submit persistent canvas operations
      description: |
        The server must authorize the participant for every batch. Candidate
        edits are rejected when candidateEditingEnabled is false; observers,
        removed participants, ended sessions, and archived sessions cannot edit.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitOperationsRequest'
      responses:
        '200':
          description: Number of accepted envelopes and the new document cursor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitOperationsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Operation rejected because the session state or cursor is no longer compatible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/sessions/{id}/canvas/export:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    get:
      tags: [Canvas]
      operationId: exportCanvasJson
      summary: Export a session and its canvas as JSON
      description: The frontend downloads this response as a JSON file.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: JSON export containing session metadata and canvas data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CanvasExport'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Signed-in user access token, or a short-lived participant collaboration
        credential returned after a guest joins. Guest-link URLs are only used
        for the initial public resolve/join exchange.
  parameters:
    SessionId:
      name: id
      in: path
      required: true
      description: Interview session identifier.
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated principal lacks permission for this operation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource does not exist or is not visible to the caller.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    TokenRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    TokenResponse:
      type: object
      required: [access_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [bearer]
        expires_in:
          type: integer
          minimum: 1
    User:
      type: object
      required: [id, email, displayName, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
        createdAt:
          $ref: '#/components/schemas/EpochMillis'
    SessionState:
      type: string
      enum: [draft, live, ended, archived]
...
    InterviewSession:
      type: object
      required:
        - id
        - ownerUserId
        - title
        - prompt
        - state
        - candidateEditingEnabled
        - durationMinutes
        - scheduledAt
        - startedAt
        - endedAt
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        ownerUserId:
          type: string
        title:
          type: string
        prompt:
          type: string
        state:
          $ref: '#/components/schemas/SessionState'
        candidateEditingEnabled:
          type: boolean
        durationMinutes:
          type: integer
          minimum: 1
        scheduledAt:
          oneOf:
            - $ref: '#/components/schemas/EpochMillis'
            - type: 'null'
        startedAt:
          oneOf:
            - $ref: '#/components/schemas/EpochMillis'
            - type: 'null'
        endedAt:
          oneOf:
            - $ref: '#/components/schemas/EpochMillis'
            - type: 'null'
        createdAt:
          $ref: '#/components/schemas/EpochMillis'
        updatedAt:
          $ref: '#/components/schemas/EpochMillis'
...
    CanvasSnapshot:
      type: object
      required: [sessionId, doc, operationCursor, updatedAt]
      properties:
        sessionId:
          type: string
        doc:
          $ref: '#/components/schemas/CanvasDoc'
        operationCursor:
          type: integer
          minimum: 0
        updatedAt:
          $ref: '#/components/schemas/EpochMillis'
    CanvasDoc:
      type: object
      required: [schemaVersion, elements, order]
      properties:
        schemaVersion:
          type: integer
          minimum: 1
        elements:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/CanvasElement'
        order:
          type: array
          description: Element IDs in draw order, from bottom to top.
          items:
            type: string
    CanvasElement:
      oneOf:
        - $ref: '#/components/schemas/NodeElement'
        - $ref: '#/components/schemas/StickyElement'
        - $ref: '#/components/schemas/TextElement'
        - $ref: '#/components/schemas/ConnectorElement'
        - $ref: '#/components/schemas/StrokeElement'
      discriminator:
        propertyName: kind
        mapping:
          node: '#/components/schemas/NodeElement'
          sticky: '#/components/schemas/StickyElement'
          text: '#/components/schemas/TextElement'
          connector: '#/components/schemas/ConnectorElement'
          stroke: '#/components/schemas/StrokeElement'
    ElementBase:
      type: object
      required: [id, createdBy, createdAt, updatedAt]
      properties:
        id:
          type: string
        createdBy:
          type: string
        createdAt:
          $ref: '#/components/schemas/EpochMillis'
        updatedAt:
          $ref: '#/components/schemas/EpochMillis'
    ElementColor:
      type: string
      enum: [neutral, accent, amber, rose]
    NodeElement:
      allOf:
        - $ref: '#/components/schemas/ElementBase'
        - type: object
          required: [kind, componentType, x, y, width, height, label, color]
          properties:
            kind:
              const: node
            componentType:
              type: string
            x:
              type: number
            y:
              type: number
            width:
              type: number
            height:
              type: number
            label:
              type: string
            description:
              type: string
            color:
              $ref: '#/components/schemas/ElementColor'
    StickyElement:
      allOf:
        - $ref: '#/components/schemas/ElementBase'
        - type: object
          required: [kind, x, y, width, height, text, color]
          properties:
            kind:
              const: sticky
            x:
              type: number
            y:
              type: number
            width:
              type: number
            height:
              type: number
            text:
              type: string
            color:
              $ref: '#/components/schemas/ElementColor'
    TextElement:
      allOf:
        - $ref: '#/components/schemas/ElementBase'
        - type: object
          required: [kind, x, y, text, size, color]
          properties:
            kind:
              const: text
            x:
              type: number
            y:
              type: number
            text:
              type: string
            size:
              type: number
            color:
              $ref: '#/components/schemas/ElementColor'
    ConnectorEndpoint:
      oneOf:
        - type: object
          required: [elementId]
          properties:
            elementId:
              type: string
        - $ref: '#/components/schemas/Point'
    ConnectorElement:
      allOf:
        - $ref: '#/components/schemas/ElementBase'
        - type: object
          required: [kind, from, to, style, arrowStart, arrowEnd, dashed, width, color]
          properties:
            kind:
              const: connector
            from:
              $ref: '#/components/schemas/ConnectorEndpoint'
            to:
              $ref: '#/components/schemas/ConnectorEndpoint'
            style:
              type: string
              enum: [straight, elbow, curved]
            arrowStart:
              type: boolean
            arrowEnd:
              type: boolean
            dashed:
              type: boolean
            width:
              type: number
            label:
              type: string
            color:
              $ref: '#/components/schemas/ElementColor'
    StrokeElement:
      allOf:
        - $ref: '#/components/schemas/ElementBase'
        - type: object
          required: [kind, points, color, width, tool]
          properties:
            kind:
              const: stroke
            points:
              type: array
              items:
                type: array
                prefixItems:
                  - type: number
                  - type: number
                minItems: 2
                maxItems: 2
            color:
              $ref: '#/components/schemas/ElementColor'
            width:
              type: number
            tool:
              type: string
              enum: [pen, highlighter]
    CanvasElementPatch:
      type: object
      description: Partial element object applied by an update operation.
      additionalProperties: true
    CanvasOperation:
      oneOf:
        - $ref: '#/components/schemas/AddOperation'
        - $ref: '#/components/schemas/UpdateOperation'
        - $ref: '#/components/schemas/DeleteOperation'
        - $ref: '#/components/schemas/ReorderOperation'
        - $ref: '#/components/schemas/ClearOperation'
    AddOperation:
      type: object
      required: [type, element]
      properties:
        type:
          const: add
        element:
          $ref: '#/components/schemas/CanvasElement'
    UpdateOperation:
      type: object
      required: [type, id, patch, at]
      properties:
        type:
          const: update
        id:
          type: string
        patch:
          $ref: '#/components/schemas/CanvasElementPatch'
        at:
          $ref: '#/components/schemas/EpochMillis'
    DeleteOperation:
      type: object
      required: [type, ids]
      properties:
        type:
          const: delete
        ids:
          type: array
          minItems: 1
          items:
            type: string
    ReorderOperation:
      type: object
      required: [type, id, to]
      properties:
        type:
          const: reorder
        id:
          type: string
        to:
          type: string
          enum: [front, back]
    ClearOperation:
      type: object
      required: [type]
      properties:
        type:
          const: clear
    CanvasEnvelope:
      type: object
      required: [clientOperationId, actorId, sentAt, operation]
      properties:
        clientOperationId:
          type: string
          description: Unique per client operation; used for deduplication.
        actorId:
          type: string
        sentAt:
          $ref: '#/components/schemas/EpochMillis'
        operation:
          $ref: '#/components/schemas/CanvasOperation'
...
    SubmitOperationsRequest:
      type: object
      required: [participantId, envelopes]
      properties:
        participantId:
          type: string
        envelopes:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CanvasEnvelope'
    SubmitOperationsResponse:
      type: object
      required: [accepted, cursor]
      properties:
        accepted:
          type: integer
          minimum: 0
        cursor:
          type: integer
          minimum: 0
    CanvasExport:
      type: object
      required: [session, canvas, exportedAt]
      properties:
        session:
          $ref: '#/components/schemas/InterviewSession'
        canvas:
          $ref: '#/components/schemas/CanvasDoc'
        exportedAt:
          type: string
          format: date-time
    ErrorCode:
      type: string
      enum: [not_found, revoked, expired, at_capacity, forbidden, ended, invalid]
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    Point:
      type: object
      required: [x, y]
      properties:
        x:
          type: number
        y:
          type: number
    EpochMillis:
      type: integer
      format: int64
      description: Unix epoch timestamp in milliseconds.
...